refactor ( #27 ) : screen 패키지 구조 수정
